pyproject.toml
[project]
name = "presto-deck-v1"
version = "0.1.0"
description = "AI-powered presentation deck generation service with event-driven architecture"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "pydantic[email,timezone]>=2.11.7",
    "pytest>=8.4.1",
    "redis>=5.0.0,<6.0.0",
    "uvicorn>=0.35.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "psycopg[binary]>=3.1.16",
    "asyncpg>=0.29.0",
    "arq>=0.26.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "bleach>=6.1.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.42b0",
    "opentelemetry-instrumentation-redis>=0.42b0",
    "opentelemetry-exporter-otlp>=1.21.0",
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    "pytest-asyncio>=0.23.2",
    "pytest-mock>=3.12.0",
    "factory-boy>=3.3.0",
    "greenlet>=3.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "opentelemetry-instrumentation-logging>=0.57b0",
    "psycopg2>=2.9.10",
    "aiosqlite>=0.21.0",
    "langchain-community>=0.3.29",
    "pre-commit>=4.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
